Reunite machines
